Update ntlmrequest.js
This helps if you are nocking the requests to the EWS server.  If you nock the requests it generates a hash using your local machine name.  Then using those auth headers, running them on a build server, the auth headers are different.  This lets you specify the hostname of the machine so the hashes are always the same and match your nocked requests.