Honour ADFS redirects.
Fixes authentication for the ADFS server I am authenticating against because it redirects both to different relative path and appends a ?SAMLRequest=... query parameter in the redirect Location. Without these subsequent stages of the authentication process fail. Setting  request.followRedirect: true isn't enough because that correctly causes redirects to be followed within a single stage of the authentication, but then subsequent stages revert to the original URL provided by the caller, not the redirected URL.